Refactor to builder pattern w/ environment variables
